\name{fitPiecewiseLinreg-methods}
\docType{methods}
\alias{fitPiecewiseLinreg-methods}
\alias{fitPiecewiseLinreg,list-method}
\title{ Methods for fitPiecewiseLinreg in flowMerge package}
\description{
 Methods for the function \code{fitPiecewiseLinreg} in the package \code{flowMerge}
}
\section{Methods}{
\describe{

\item{x = "list"}{ A list of \code{flowMerge} objectes derived from a call to the \code{merge} function. }
}}
\references{ Finak G, Bashasharti A, Brinkmann R, Gottardo R. Merging Mixture Model Components for Improved Cell Population Identification in High Throughput Flow Cytometry Data (Submitted)}
\keyword{methods}
