\name{flagOutliers-methods}
\docType{methods}
\alias{flagOutliers-methods}
\alias{flagOutliers,flowMerge-method}
\title{ Methods to update the flagOutliers slot in a flowMerge object. }
\description{
 Methods that update the \code{flagOutliers} slot in a flowMerge object so that they reflec the outliers in the new merged clustering. This is an internal function, not meant for user consumption. It is called from within the \code{merge} method.
}
\section{Methods}{
\describe{
\item{object = "flowMerge"}{ Update the \code{flagOutliers} slot for an object of type \code{flowMerge} }
}}
\keyword{methods}
