\name{flagOutliers}
\alias{flagOutliers}
\title{ Update the flagOutliers slot in a flowMerge object }
\description{
  Update the flagOutliers slot in a flowMerge object. This method is internal and called automatically from within the merging code.
}
\usage{flagOutliers(object,...)}
\arguments{
  \item{object}{ An object of type \code{flowMerge} }
  \item{...}{Additional arguments, currently unused}.
}

%\references{
%}
\keyword{misc}
