\name{flowMerge-internal}
\title{Internal flowMerge functions}
\alias{BIC}
\alias{ENT}
\alias{ICL}
\description{
Internal functions for extracting the BIC, ICL or entropy from a list of flowClus, flowObj, or flowMerge objects.
}
\references{ Finak G, Bashasharti A, Brinkmann R, Gottardo R. Merging Mixture Model Components for Improved Cell Population Identification in High Throughput Flow Cytometry Data (Submitted)}
\keyword{internal}
