\name{mergeClusters}
\alias{mergeClusters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Cluster merging not meant to be called by the user }
\description{
  Internal cluster merging function.
}
\usage{
mergeClusters(object, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ not meant to be called by the user }
  \item{data}{ not meant to be called by the user }
}
\details{
  Not meant to be called by the user
}
\value{
  Not meant to be called by the user
  }
\references{ Finak G, Bashasharti A, Brinkmann R, Gottardo R. Merging Mixture Model Components for Improved Cell Population Identification in High Throughput Flow Cytometry Data (Submitted)}
\author{Greg Finak \email{<greg.finak@ircm.qc.ca>} }

\seealso{ \link{merge} }
\keyword{ cluster }

