\name{mergeClusters2}
\alias{mergeClusters2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Cluster merging not meant to be called by the user }
\description{
  Internal function not meant to be called by the user.
}
\usage{
mergeClusters2(object, a, b, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ Internal function not meant to be called by the user. }
  \item{a}{ Internal function not meant to be called by the user. }
  \item{b}{ Internal function not meant to be called by the user. }
  \item{data}{ Internal function not meant to be called by the user. }
}
\details{
  Internal function not meant to be called by the user.
}
\value{
  Internal function not meant to be called by the user.}
\references{ Finak G, Bashasharti A, Brinkmann R, Gottardo R. Merging Mixture Model Components for Improved Cell Population Identification in High Throughput Flow Cytometry Data (Submitted)}
\author{ Greg Finak \email{<greg.finak@ircm.qc.ca>} }
\seealso{ \link{merge} }
\keyword{ cluster}

