\name{show-methods}
\docType{methods}
\alias{show,flowMerge-method}
\alias{show,flowObj-method}
\title{ Describe a flowObj or flowMerge object}
\description{
 Accessors to describe a \code{flowObj} or \code{flowMerge} object.
}
\section{Methods}{
\describe{

\item{object = "flowMerge"}{ Describe a \code{flowMerge} object. }

\item{object = "flowObj"}{ Describe a \code{flowObj} object. }
}}
\keyword{methods}
