\name{summary-methods}
\docType{methods}
\alias{summary,flowMerge-method}
\alias{summary,flowObj-method}
\title{Summary methods for flowMerge}
\description{
 Summary method for \code{flowMerge} objects.
}
\section{Methods}{
\describe{

\item{object = "flowMerge"}{ Summarize a flowMerge object. }

\item{object = "flowObj"}{ Summarize a flowObj object }
}}
\keyword{methods}
