\name{split-methods}
\docType{methods}
\alias{split-methods}
\alias{split,missing,flowMerge-method}
\title{ Split data in a flowMerge object by cluster}
\description{
 Split method defined for flowMerge objects. Pulls out the population based on cluster number.
}
\section{Methods}{
\describe{

\\item{x = "missing", f = "flowMerge"}{ Split a flowMerge object into its component clusters. }
}}
\keyword{methods}
